requirements.txt
defusedxml==0.7.1
lml==0.1.0
odfpy==1.4.1
pyexcel-io==0.6.6
pyexcel-ods==0.6.0
python-dotenv==1.0.0
